+++ Octokit.Reactive/Clients/IObservableOrganizationHooksClient.cs
@@ -28,7 +28,7 @@ namespace Octokit.Reactive
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#get-single-hook">API documentation</a> for more information.</remarks>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get", Justification = "This is ok; we're matching HTTP verbs not keyworks")]
-        IObservable<OrganizationHook> Get(string org, int hookId);
+        IObservable<OrganizationHook> Get(string org, long hookId);
 
         /// <summary>
         /// Creates a hook for a organization
@@ -44,7 +44,7 @@ namespace Octokit.Reactive
         /// <param name="hookId">The organizations hook id</param>
         /// <param name="hook">The hook's parameters</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
-        IObservable<OrganizationHook> Edit(string org, int hookId, EditOrganizationHook hook);
+        IObservable<OrganizationHook> Edit(string org, long hookId, EditOrganizationHook hook);
 
         /// <summary>
         /// This will trigger a ping event to be sent to the hook.
@@ -52,7 +52,7 @@ namespace Octokit.Reactive
         /// <param name="org">The organizations name</param>
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#ping-a-hook">API documentation</a> for more information.</remarks>
-        IObservable<Unit> Ping(string org, int hookId);
+        IObservable<Unit> Ping(string org, long hookId);
 
         /// <summary>
         /// Deletes a hook for a organization
@@ -60,6 +60,6 @@ namespace Octokit.Reactive
         /// <param name="org">The organizations name</param>
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="https://developer.github.com/v3/orgs/hooks/#delete-a-hook">API documentation</a> for more information.</remarks>
-        IObservable<Unit> Delete(string org, int hookId);
+        IObservable<Unit> Delete(string org, long hookId);
     }
 }
