+++ Octokit.Reactive/Clients/ObservableOrganizationHooksClient.cs
@@ -50,7 +50,7 @@ namespace Octokit.Reactive
         /// <param name="org">The organizations name</param>
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/orgs/hooks/#get-single-hook">API documentation</a> for more information.</remarks>
-        public IObservable<OrganizationHook> Get(string org, int hookId)
+        public IObservable<OrganizationHook> Get(string org, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
 
@@ -79,7 +79,7 @@ namespace Octokit.Reactive
         /// <param name="hook">The hook's parameters</param>
         /// <remarks>See <a href="http://developer.github.com/v3/orgs/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
         /// <returns></returns>
-        public IObservable<OrganizationHook> Edit(string org, int hookId, EditOrganizationHook hook)
+        public IObservable<OrganizationHook> Edit(string org, long hookId, EditOrganizationHook hook)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
             Ensure.ArgumentNotNull(hook, nameof(hook));
@@ -93,7 +93,7 @@ namespace Octokit.Reactive
         /// <param name="org">The organizations name</param>
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/orgs/hooks/#ping-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<Unit> Ping(string org, int hookId)
+        public IObservable<Unit> Ping(string org, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
 
@@ -106,7 +106,7 @@ namespace Octokit.Reactive
         /// <param name="org">The organizations name</param>
         /// <param name="hookId">The organizations hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/orgs/hooks/#delete-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<Unit> Delete(string org, int hookId)
+        public IObservable<Unit> Delete(string org, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(org, nameof(org));
